+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project, at least loosely, adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- model.phase() now defaults to abs_phase=True when TZR* params are in the model
 ### Added
 - logging now needs to be setup explicitly
 - Color-by-jump mode for pintk